pyproject.toml
[project]
name = "imagegen"
version = "0.1.0"
description = "CLI tool for generating images using NanoBanana API providers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "google-genai>=1.0.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
imagegen = "imagegen.cli:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12"]
build-backend = "uv_build"

[dependency-groups]
dev = ["ruff>=0.8.0", "mypy>=1.13.0"]
test = ["pytest>=8.0.0"]
